Moving ContentGroup component to global scope as requested in PR review
