+++ sources/packages/web/src/components/ContentGroup.vue
